Remove classes, use functions
